Updated examples

+++ Classes/GHTest/GHTesting.m
@@ -62,7 +62,7 @@ NSInteger ClassSort(id a, id b, void *context) {
 
 // GTM_BEGIN
 // Used for sorting methods below
-static int MethodSort(id a, id b, void *context) {
+static NSInteger MethodSort(id a, id b, void *context) {
   NSInvocation *invocationA = a;
   NSInvocation *invocationB = b;
   const char *nameA = sel_getName([invocationA selector]);
